CDK-501: Add more view URI tests.
This uses local file dataset URIs to test, but the test applies to any
implementation that uses AbstractRefinableView. Introducing a test
dataset, view, repository, and URI pattern to make this slightly more
generic wasn't worth the work.

This fixes URIBuilder; when constraints are empty, the URI should be a
dataset URI. Also fixes AbstractDataset, which should implement equals
and hashCode to make these tests work.