CDK-549: Add DoFn transform to copy and import commands.
This adds --jar and --transform options to the copy and import commands.
The jar option will add a jar to the distributed cache and the current
classloader. The transform option is the name of a DoFn class that will
be instantiated and used to transform GenericRecord objects. It must
produce GenericRecord objects.