CDK-549: Use new ClassLoader for user-supplied transforms.
Before this, the addJars method was invoking addURL on the current
loader, which is a hidden method and isn't allowed on secure JVMs. This
commit uses the ClassLoader handling from obj-schema to load the
user-supplied transform class and create a TransformTask. This also adds
a addJars method to the TaskUtil configuration helper.