CDK-140: Add backward-compatibility to FS datasets.
This adds backward-compatibility to the FS metadata provider. There is
no need for more backward-compatible checks as long as the metadata is
found because the location is set when a dataset is created and will
always be returned from the descriptor metadata.

The FileSystemMetadataProvider will now check for datasets located in
the provider's root folder and consider them part of the default
namespace. Tests are included.

This compatibility is needed so that callers do not have to change the
repository API. Existing configurations for the kite-maven-plugin and
DatasetSink should continue to work.

Conflicts:
	kite-data/kite-data-core/src/main/java/org/kitesdk/data/spi/filesystem/FileSystemMetadataProvider.java