CDK-653: Add DefaultConfiguration.
This adds DefaultConfiguration, which is a single place to get the
environment Hadoop configuration. Previously, the Loader classes would
call new Configuration() to inspect the environment and get defaults,
like the HDFS location. Now, the loaders call DefaultConfiguration.get()
instead that returns a consistent Configuration to all callers.

There is also a DefaultConfiguration.set(Configuration) method that can
replace the default Configuration based on just the environment. This
should only be used to replace the environment config with a more
accurate global configuration. For example, one configured with CLI
options by ToolRunner or via other configuration methods that apply
globally to the process.